requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
djangorestframework==3.14.0
dj-rest-auth==4.0.1
django-extensions==3.2.3
django-cors-headers==4.2.0